[12기 문상호] Add : 컴포넌트 분리 및 상품등록/기본정보/카테고리 기능 구현
