+++ public/Data/ProductAdd/mockData.json
@@ -2,7 +2,9 @@
     "data": {
         "product_add": {
             "category": {
+                "1차 카테고리를 선택해주세요": ["1차 카테고리를 먼저 선택해주세요"],
                 "아우터" : [
+                    "2차 카테고리를 선택해주세요",
                     "자켓",
                     "가디건",
                     "코트",
@@ -12,6 +14,7 @@
                     "기타"
                 ],
                 "상의": [
+                    "2차 카테고리를 선택해주세요",
                     "티셔츠",
                     "셔츠/블라우스",
                     "니트",
@@ -19,6 +22,7 @@
                     "베스트"
                 ],
                 "바지": [
+                    "2차 카테고리를 선택해주세요",
                     "청바지",
                     "슬랙스",
                     "면바지",
@@ -27,6 +31,7 @@
                     "레깅스"
                 ],
                 "원피스": [
+                    "2차 카테고리를 선택해주세요",
                     "미니",
                     "미디",
                     "롱",
@@ -34,11 +39,13 @@
                     "점프수트"
                 ],
                 "스커트": [
+                    "2차 카테고리를 선택해주세요",
                     "미니",
                     "미디",
                     "롱"
                 ],
                 "신발": [
+                    "2차 카테고리를 선택해주세요",
                     "플랫/로퍼",
                     "샌들/슬리퍼",
                     "힐",
@@ -46,6 +53,7 @@
                     "부츠/워커"
                 ],
                 "가방": [
+                    "2차 카테고리를 선택해주세요",
                     "크로스백",
                     "토트백",
                     "숄더백",
@@ -54,12 +62,14 @@
                     "백팩"
                 ],
                 "주얼리": [
+                    "2차 카테고리를 선택해주세요",
                     "귀걸이",
                     "목걸이",
                     "팔찌/발찌",
                     "반지"
                 ],
                 "잡화": [
+                    "2차 카테고리를 선택해주세요",
                     "휴대폰 acc",
                     "헤어 acc",
                     "양말/스타킹",
@@ -72,6 +82,7 @@
                     "기타"
                 ],
                 "라이프웨어": [
+                    "2차 카테고리를 선택해주세요",
                     "언더웨어",
                     "홈웨어",
                     "스윔웨어",
@@ -79,6 +90,7 @@
                     "기타"
                 ],
                 "빅사이즈": [
+                    "2차 카테고리를 선택해주세요",
                     "아우터",
                     "상의",
                     "바지",
